+++ esp-idf-svc/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add option to specify initial caps for the MQTT async adaptor vectors
 - Allow using esp timer with skip_unhandled_events (#526)
 - OTA - Implements a new type `EspFirmwareInfoLoad` that has a reduced memory consumption (#531)
+- Added set_promiscuous function in EthDriver (#246)
 
 ### Fixed
 - The alloc::Vec version stomps the scan state from Done to Idle. (#459)